POCO Persist with files
Added the server folder "Surveys" and in that added 3 json files with survey definitions. The first and second are copies from the original demo.
The third is for demonstration
You will need to track json data when send between client and server. Then you can copy the json content into a physical file which will be read at next session startup.
Files are only read at startup. Changes and new surveys are kept in session storage. Track the json survey definition and save them to a physical file